+++ src/main/java/com/researchspace/chemistry/convert/convertor/OpenBabelConvertor.java
@@ -1,6 +1,6 @@
 package com.researchspace.chemistry.convert.convertor;
 
-import com.researchspace.chemistry.ChemistryException;
+import com.researchspace.chemistry.exception.ChemistryException;
 import com.researchspace.chemistry.convert.ConvertDTO;
 import com.researchspace.chemistry.util.CommandExecutor;
 import jakarta.annotation.PostConstruct;
