+++ src/main/java/com/researchspace/chemistry/image/ImageService.java
@@ -1,6 +1,6 @@
 package com.researchspace.chemistry.image;
 
-import com.researchspace.chemistry.ChemistryException;
+import com.researchspace.chemistry.exception.ChemistryException;
 import com.researchspace.chemistry.image.generator.IndigoImageGenerator;
 import com.researchspace.chemistry.image.generator.OpenBabelImageGenerator;
 import java.util.Optional;
