+++ src/main/java/com/researchspace/chemistry/util/IndigoFacade.java
@@ -3,7 +3,7 @@ package com.researchspace.chemistry.util;
 import com.epam.indigo.Indigo;
 import com.epam.indigo.IndigoException;
 import com.epam.indigo.IndigoObject;
-import com.researchspace.chemistry.ChemistryException;
+import com.researchspace.chemistry.exception.ChemistryException;
 import com.researchspace.chemistry.convert.ConvertDTO;
 import java.util.Optional;
 import org.apache.commons.lang3.StringUtils;
