+++ src/test/java/com/researchspace/chemistry/analysis/IndigoChemicalAnalyserTest.java
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.researchspace.chemistry.ChemistryException;
+import com.researchspace.chemistry.exception.ChemistryException;
 import com.researchspace.chemistry.extract.ExtractionResult;
 import java.util.List;
 import org.junit.jupiter.api.Test;
