+++ src/test/java/com/researchspace/chemistry/convert/ConvertServiceIT.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.researchspace.chemistry.ChemistryException;
+import com.researchspace.chemistry.exception.ChemistryException;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
