+++ src/test/java/com/researchspace/chemistry/convert/convertor/OpenBabelConvertorTest.java
@@ -5,7 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
-import com.researchspace.chemistry.ChemistryException;
+import com.researchspace.chemistry.exception.ChemistryException;
 import com.researchspace.chemistry.convert.ConvertDTO;
 import com.researchspace.chemistry.util.CommandExecutor;
 import java.io.IOException;
